I made register work
