+++ RestaurantChapeau/App.config
@@ -1,7 +1,7 @@
 ﻿<?xml version="1.0" encoding="utf-8" ?>
 <configuration>
 	<connectionStrings>
-		<add name="SomerenDatabase" connectionString="Data Source=;
-         Initial Catalog=; User=; Password="/>
+		<add name="ChapeauDatabase" connectionString="Data Source=2122chapeau.database.windows.net; Initial Catalog=db_IT1b4
+; User=group_IT1b4; Password='h4$%juH-eUQq'"/>
 	</connectionStrings>
 </configuration>
