+++ RestaurantChapeau/ErrorLog.txt
@@ -1,0 +1,10 @@
+Time: 20/05/2022 10:37:05 am
+-----------------------------------------------------------
+Message: There is no user with these credentials
+-----------------------------------------------------------
+
+Time: 20/05/2022 10:37:59 am
+-----------------------------------------------------------
+Message: There is no user with these credentials
+-----------------------------------------------------------
+
