+++ RestaurantChapeau/Program.cs
@@ -17,7 +17,7 @@ namespace RestaurantChapeau
             Application.SetHighDpiMode(HighDpiMode.SystemAware);
             Application.EnableVisualStyles();
             Application.SetCompatibleTextRenderingDefault(false);
-            Application.Run(new Form1());
+            Application.Run(new LoginForm());
         }
     }
 }
